Progress towards our dotmobile.us being able to handle cert generation, storage, and retreival (needed for the initial login to work correctly)
Right now, our CA is stored with an unencrypted private key. As we are not actually trusting these certs right now, I'm not sure its a problem


For the initial setup, the user must run dotmacCAsetup.pl, and then change the ownership of the certs directory to the www-data
user.

there are a couple dotmac.conf revisions that will be in the next checkin.

The client portion of this is partially handled by security_dotmac_tp, which is available on www.opensource.apple.com

-Rob